# Optimize initial `Node::get_path` call by avoiding `Vector::reverse`